Delivering HOPS 3.23 vendor branch at Haystack svn rev 3383 from tarball (Oct 14 03:00)/swc/tarpit/hops-tc-3.23swc.tar.gz
changes include corrections to normalization process for passband
and notches mode and some primitive ability to make the cross-power
spectrum plot more useful.  Minor changes to aedit to be more helpful
when required files are not found.  chops mk4_module introduced.
Other changes required by the HOPS4 project (hops_complex.h).



git-svn-id: https://svn.atnf.csiro.au/difx@10257 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/hops/trunk/chops/source/c_src/fourfit_libs/ffcontrol/src/read_control_file.c
@@ -54,7 +54,7 @@ int read_control_file (char* control_file_name, char** input_string, int* flag)
    num_chars = 0;
    num_line = 1;
 
-   if (memcmp (control_file_name, "if ", 3))   /* special memory file option? */
+   if (memcmp (control_file_name, "if ", (size_t)3))   /* special memory file option? */
       {
       *flag = IS_CONTROL_FILE;
       n = -1;                                  /* no, signify input from file */
