+++ applications/hops/trunk/chops/source/c_src/fourfit_libs/ffio/src/fill_205.c
@@ -27,7 +27,7 @@ struct type_205 *t205)
     int i, j, ch, nch, int_time, sb, ind,
         nchan, nfreqs;
     struct freq_corel *fc;
-    nchan = (strncmp (t203->version_no, "00", 2) == 0) ? 32 : 8*MAXFREQ;
+    nchan = (strncmp (t203->version_no, "00", (size_t)2) == 0) ? 32 : 8*MAXFREQ;
 
     clear_205 (t205);
                                         /* For now, UCT central is same as FRT */
